requirements.txt
rich>=13.7.0
requests>=2.31.0
beautifulsoup4==4.12.3
keyboard==0.13.5
openpyxl==3.1.2
faker==24.9.0
aiosmtplib>=2.0.0
cachetools>=5.0.0
python-dotenv>=1.0.0
questionary>=2.0.1
Pillow>=11.1.0
colorama>=0.4.6
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
